Switch to float

+++ src/main/scala/mllab/evaluaton/Evaluation.scala
@@ -34,6 +34,6 @@ object Evaluation{
     println("F " + falsePositives + " " + falseNegatives)
   }
 
-  def MSE(y_pred: List[Double], y_true: List[Double]): Float =
+  def MSE(y_pred: List[Float], y_true: List[Float]): Float =
     (y_pred zip y_true).map{case (x, y) => Math.pow(x - y, 2)}.sum.toFloat
 }
